Fix unmounting bug for real
